move sidbar files into navigation folder
