+++ src/App.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react'
 import { useSelector } from 'react-redux'
 import './styles/globals.css'
-import Sidebar from './components/Sidebar'
+import Sidebar from './components/navigation/Sidebar'
 import AdminDashboard from './views/AdminDashboard'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
-import PhoneSidebar from './components/PhoneSidebar'
+import PhoneSidebar from './components/navigation/PhoneSidebar'
 import LoginPage from './components/LoginPage'
 import ErrorSnackbar from './components/ErrorSnackbar'
 //import Authentication from './modules/Authentication'
