+++ src/components/navigation/PhoneSidebar.jsx
@@ -4,14 +4,15 @@ import List from '@material-ui/core/List'
 import Divider from '@material-ui/core/Divider'
 import IconButton from '@material-ui/core/IconButton'
 import MailIcon from '@material-ui/icons/Mail'
-import SidebarIcon from './SidebarIcon'
 import MenuIcon from '@material-ui/icons/Menu'
-import logoWhite from '../assets/LogoCHWLSymbolWhite.png'
 import AssessmentIcon from '@material-ui/icons/Assessment'
 import ExitToAppIcon from '@material-ui/icons/ExitToApp'
-import Authentication from '../modules/Authentication'
 import { Grid, makeStyles } from '@material-ui/core'
 import ArticleIcon from '@material-ui/icons/Subject'
+
+import SidebarIcon from './SidebarIcon'
+import logoWhite from '../../assets/LogoCHWLSymbolWhite.png'
+import Authentication from '../../modules/Authentication'
 
 const useStyles = makeStyles((theme) => ({
   menuBar: {
